Add -V flag (unvendor) to strip */vendor/ prefixes
This ough to ignore the vendor packages in the GOROOT, but this isn't
thoroughly tested yet.

This breaks pkgs out into their canonical import paths -- i.e., without
vendor import paths. However, it will still process all different
versions of a vendored path, so that might look a little weird in the
long run.

This also fixes the srcDir when processing packages, as it will no
longer attempt to import packages from a singular root. This makes it
possible to process vendor packages a little better. Not fully tested,
either.

Change-Id: Ie4e138eb2b99325c7b7dad4525a9bbc3e6d24e32